pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "two-minds-one-world"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.30.0",
    "celery[redis]>=5.3",
    "psycopg2-binary>=2.9",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "duckduckgo-search>=6.0",
    "requests>=2.31",
    "html2text>=2024.2",
    "playwright>=1.40.0",
    "browser-use>=0.1.40",
    "langchain-openai>=0.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "tools*", "db*", "world*", "config*"]
